app/requirements.txt
django==4.2
djangorestframework==3.14.0
djoser==2.2.0
